Zmiana nazwy
Zmiana nazwy z Mori — kopia.py na Mori.py
